Revert "WIP: move to OffscreenCanvas"
This reverts commit 5b4a9cf6634b0bb7d9cf344b897dcad320f14bf7.
