--- packages/scan/src/canvas/store.ts
@@ -1,9 +1,0 @@
-import type { ActiveOutline } from "./types";
-
-interface Store {
-	activeOutlines: Array<ActiveOutline>;
-}
-
-export const STORE: Store = {
-	activeOutlines: [],
-};
